backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
google-generativeai==0.3.2
python-docx==1.0.1
PyPDF2==3.0.1
pandas==2.1.4
Werkzeug==3.0.1 